FDG-6631 moved to helper
